Django/requirements.txt
Django==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1